requirements.txt
# Telegram bot
aiogram==3.10.0
nest_asyncio
python-dotenv

# Database
SQLAlchemy==2.0.35
psycopg2-binary  # если PostgreSQL

# Server / API
fastapi
uvicorn

# HTTP-запросы
requests

# Gemini NEW API client
google-genai>=0.3.0
